Update index.js
Recently learned about the difference between .innerText and .textContent. 
innerText will ignore blank spaces where as .textContent will not so it was better to to change the .innerText on both sides. 

Also was able to follow along this YouTube video to be able to make the save button also reset the number that is shown